refactor: simplify file naming for twMerge
